+++ Scripts/main.js
@@ -1,79 +1,27 @@
+import displayBox from './Modules/displayBox.js';
+import Intialize from './Modules/initial.js';
+import myStorage from './Modules/myStorage.js';
+
 const getAppended = document.getElementById('book-container');
-class AwesomeBooks {
-  constructor(title, author, id) {
-    this.title = title;
-    this.author = author;
-    this.id = id;
-  }
 
-  static getData() {
-    let books;
-    if (localStorage.getItem('books') === null) {
-      books = [];
-    } else {
-      books = JSON.parse(localStorage.getItem('books'));
-    }
-    return books;
-  }
-
-  static addition(book) {
-    const books = AwesomeBooks.getData();
-    books.push(book);
-    localStorage.setItem('books', JSON.stringify(books));
-  }
-
-  static remove(id) {
-    const books = AwesomeBooks.getData();
-    books.forEach((book, index) => {
-      if (book.id === id) {
-        books.splice(index, 1);
-      }
-    });
-    localStorage.setItem('books', JSON.stringify(books));
-  }
-
-  static displayBooks() {
-    const books = AwesomeBooks.getData();
-    const myBooks = books;
-    myBooks.forEach((book) => {
-      AwesomeBooks.addBookToList(book);
-    });
-  }
-
-  static addBookToList(book) {
-    const beAppended = document.createElement('div');
-    beAppended.className = 'kitab';
-    beAppended.innerHTML = `
-       <h4>${book.title}</h4> <span class = 'creator'>${book.author}</span>
-       <p class="identity-book">${book.id}</p>
-       <button class = 'delete'>Delete</button>
-       `;
-    getAppended.appendChild(beAppended);
-  }
-
-  static clearFields() {
-    document.getElementById('title').value = '';
-    document.getElementById('author').value = '';
-  }
-}
-document.addEventListener('DOMContentLoaded', AwesomeBooks.displayBooks());
+document.addEventListener('DOMContentLoaded', displayBox.displayBooks());
 document.getElementById('form').addEventListener('submit', (e) => {
   const title = document.getElementById('title').value;
   const author = document.getElementById('author').value;
   const id = Math.floor(Math.random() * 1000) + 1;
   id.toString();
   e.preventDefault();
-  const newBook = new AwesomeBooks(title, author, id);
-  AwesomeBooks.addBookToList(newBook);
-  AwesomeBooks.clearFields();
-  AwesomeBooks.addition(newBook);
+  const newBook = new Intialize(title, author, id);
+  displayBox.addBookToList(newBook);
+  displayBox.clearFields();
+  myStorage.addition(newBook);
 });
 getAppended.addEventListener('click', (e) => {
   if (e.target.classList.contains('delete')) {
     e.target.parentElement.remove();
   }
   const tity = parseInt(e.target.parentElement.children[2].innerText, 10);
-  AwesomeBooks.remove(tity);
+  myStorage.remove(tity);
 });
 const getList = document.querySelector('.list');
 const getAddNew = document.querySelector('.add_new');
