+++ index.html
@@ -49,7 +49,6 @@
     <footer id="footer">
         <p>&copy; Copyright 2022</p>
     </footer>
-    <script src="Scripts/book.js"></script>
+    <script type="module" src="Scripts/main.js"></script>
 </body>
-
 </html>
